Replace occurences of ratsnest wires to ratsnest lines.
The term ratsnest wire sometimes is mistaken as a phyical connection.
We already use 'ratsnest line' at other occasions.
For translators: If a direct translation doesn't have the sound to it,
a ratsnest line could also be seen as one of the following:
net line
logical connection
unrouted connection
rubber band (connection)
Fritzing code uses the class 'VirtualWire'.
Maybe avoid terms like "air wire". This could be mixed up with
"fly wires", which are real wires soldered on a PCB.